Fix broken/misdirected "stop" command
There was a misunderstanding/misinterpretation of the OI commands associated with the word "Stop".  Opcode 173 stops the OI.  Opcode 133 actually stops a cleaning run by powering down the unit (which just puts it to sleep).  I this case the stop button on the HA interface was triggering a code 133, not good.  

With this change, the interface now directs it to behave appropriately.